+++ src/main/java/gift/domain/cart/CartItemController.java
@@ -2,7 +2,7 @@ package gift.domain.cart;
 
 import gift.domain.product.Product;
 import gift.domain.user.dto.UserInfo;
-import gift.global.jwt.JwtAuthorization;
+import gift.global.resolver.LoginInfo;
 import gift.global.response.ResponseMaker;
 import gift.global.response.ResultResponseDto;
 import gift.global.response.SimpleResultResponseDto;
@@ -36,10 +36,9 @@ public class CartItemController {
      */
     @PostMapping("/cart/{id}")
     public ResponseEntity<SimpleResultResponseDto> addCartItem(
-        @PathVariable("id") Long productId, HttpServletRequest request) {
+        @PathVariable("id") Long productId, @LoginInfo UserInfo userInfo) {
 
-        Long userId = (Long) request.getAttribute("id");
-        cartItemService.addCartItem(userId, productId);
+        cartItemService.addCartItem(userInfo.getId(), productId);
 
         return ResponseMaker.createSimpleResponse(HttpStatus.OK, "상품이 장바구니에 추가되었습니다.");
     }
@@ -49,11 +48,9 @@ public class CartItemController {
      */
     @GetMapping("/cart")
     public ResponseEntity<ResultResponseDto<List<Product>>> getProductsInCartByUserId(
-        HttpServletRequest request) {
+        @LoginInfo UserInfo userInfo) {
 
-        Long userId =   (Long) request.getAttribute("id");
-        System.out.println("userId = " + userId);
-        List<Product> products = cartItemService.getProductsInCartByUserId(userId);
+        List<Product> products = cartItemService.getProductsInCartByUserId(userInfo.getId());
 
         return ResponseMaker.createResponse(HttpStatus.OK, "장바구니 조회에 성공했습니다.", products);
     }
@@ -63,10 +60,9 @@ public class CartItemController {
      */
     @DeleteMapping("/cart/{id}")
     public ResponseEntity<SimpleResultResponseDto> deleteCartItem(
-        @PathVariable("id") Long productId, HttpServletRequest request) {
+        @PathVariable("id") Long productId, @LoginInfo UserInfo userInfo) {
 
-        Long userId = (Long) request.getAttribute("id");
-        cartItemService.deleteCartItem(userId, productId);
+        cartItemService.deleteCartItem(userInfo.getId(), productId);
 
         return ResponseMaker.createSimpleResponse(HttpStatus.OK, "장바구니에서 상품이 삭제되었습니다.");
     }
