+++ src/main/java/gift/domain/product/ProductController.java
@@ -1,7 +1,5 @@
 package gift.domain.product;
 
-import gift.domain.user.dto.UserInfo;
-import gift.global.jwt.JwtAuthorization;
 import gift.global.response.ResponseMaker;
 import gift.global.response.ResultResponseDto;
 import gift.global.response.SimpleResultResponseDto;
